modify resource view bug
